pour OWEN

+++ XOR/Maths.h
@@ -1,7 +1,7 @@
 #ifndef MATHS_H
 #define MATHS_H
 
-double sigmoid(double x);
-double der_sigmoid(double x);
+float sigmoid(float x);
+float der_sigmoid(float x);
 
 #endif
